+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - Enable inherit_errexit; require Bash 5.1+
+- Bash + zsh tab completion from the doc-block grammar
 
 ### Changed